+++ source/usbloader/wbfs.h
@@ -35,7 +35,7 @@ extern "C"
     s32 WBFS_DiskSpace(f32 *, f32 *);
     s32 WBFS_RenameGame(u8 *, const void *);
     s32 WBFS_ReIDGame(u8 *discid, const void *newID);
-    f32 WBFS_EstimeGameSize(void);
+    u64 WBFS_EstimeGameSize(void);
 
     int WBFS_GetFragList(u8 *id);
