+++ source/usbloader/wbfs/wbfs_base.h
@@ -34,7 +34,7 @@ class Wbfs
         virtual s32 DiskSpace(f32 *, f32 *) = 0;
         virtual s32 RenameGame(u8 *, const void *) = 0;
         virtual s32 ReIDGame(u8 *discid, const void *newID) = 0;
-        virtual f32 EstimateGameSize(void) = 0;
+        virtual u64 EstimateGameSize(void) = 0;
     protected:
         static u32 nb_sectors;
         wbfs_t *hdd;
